So much dememulation

+++ src/main/java/legend/game/combat/environment/BattleStageDarkening1800.java
@@ -2,6 +2,6 @@ package legend.game.combat.environment;
 
 /** Used to darken the stage during things like counterattacks and dragoon transformations */
 public class BattleStageDarkening1800 {
-  public final short[][] original_000 = new short[16][64];
-  public final short[][] modified_800 = new short[16][64];
+  public final int[][] original_000 = new int[16][64];
+  public final int[][] modified_800 = new int[16][64];
 }
